requirements.txt
# Air Quality Monitoring & Prediction — Python Dependencies
# Install:  pip install -r requirements.txt
# Python:   3.9+

# ── Core data science ──────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# ── Machine learning ───────────────────────────────────────────────────────
scikit-learn>=1.3.0

# ── HTTP & scheduling ──────────────────────────────────────────────────────
requests>=2.31.0
APScheduler>=3.10.4

# ── Configuration ──────────────────────────────────────────────────────────
PyYAML>=6.0.1

# ── Streamlit & extensions ─────────────────────────────────────────────────
streamlit>=1.35.0
streamlit-autorefresh>=1.0.1

# ── Visualisation ──────────────────────────────────────────────────────────
plotly>=5.18.0

# ── Database ───────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.0   # PostgreSQL driver for Supabase

# ── Utilities ──────────────────────────────────────────────────────────────
python-dateutil>=2.8.2
